backend - key vault/requirements.txt
flask
psycopg2-binary
flask-cors
azure-identity
azure-keyvault-secrets